split component in 2
